% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence_functions.R
\name{create_sentence_filtered_df}
\alias{create_sentence_filtered_df}
\title{Returns a filtered
  (full text) df based on the returned indices}
\usage{
create_sentence_filtered_df(df, filter_pattern, search_pattern, window)
}
\description{
Returns a filtered
  (full text) df based on the returned indices
}
