% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cx_extra_functions.R
\name{cx_extra_chart}
\alias{cx_extra_chart}
\title{Count hits/chart}
\usage{
cx_extra_chart(
  extra_input_phrase,
  df_dok,
  df_modus,
  case_sensitive,
  modus,
  indices_included = NULL
)
}
\arguments{
\item{extra_input_phrase}{phrase from extra chart field}

\item{df_dok}{data_dok df}

\item{df_modus}{data_dok or data_365 df}

\item{case_sensitive}{search_arguments$case_sensitive}

\item{modus}{plot_mode}

\item{indices_included}{R data_dok IDs included from potential filtering before this function
is called}
}
\value{
base R data frame with nrow() == nrow(df_modus) and 1 column with hits per row
}
\description{
Chart hits on corpus based on one "extra chart field" phrase
}
