% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cx_extra_functions.R
\name{cx_extra_create_df_for_info}
\alias{cx_extra_create_df_for_info}
\title{"Extra" equivalent to create_df_for_info() in corporaexplorer}
\usage{
cx_extra_create_df_for_info(df, terms, case_sensitive)
}
\arguments{
\item{df}{session_variables$data_dok,}

\item{terms}{search_arguments$extra_chart_terms}

\item{case_sensitive}{search_arguments$case_sensitive}
}
\value{
A peculiar list to be used to create table and
  plot in corpus info tab
}
\description{
"Extra" equivalent to create_df_for_info() in corporaexplorer
}
