% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cx_extra_functions.R
\name{cx_extra_reset_data}
\alias{cx_extra_reset_data}
\title{Actions to be carried out when search button is clicked}
\usage{
cx_extra_reset_data()
}
\description{
Actions to be carried out when search button is clicked
}
