% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cx_extra_functions.R
\name{cx_extra_subset}
\alias{cx_extra_subset}
\title{Subset corpus}
\usage{
cx_extra_subset(
  extra_input_phrase,
  df,
  case_sensitive,
  indices_included = NULL
)
}
\arguments{
\item{extra_input_phrase}{phrase from extra subset field}

\item{df}{data_dok corporaexplorer data frame}

\item{case_sensitive}{search_arguments$case_sensitive}

\item{indices_included}{R data_dok IDs included from potential filtering before this function
is called}
}
\value{
filtered data_dok corporaexplorer data frame
}
\description{
Filter corpus based on one "extra subset field" phrase
}
