% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cx_extra_functions.R
\name{cx_extra_tab_text}
\alias{cx_extra_tab_text}
\title{Text to be displayed in "extra tab" in document box}
\usage{
cx_extra_tab_text(df, min_rad, patterns)
}
\arguments{
\item{df}{corporaexplorer df (data_dok or data_365)}

\item{min_rad}{row in df}

\item{patterns}{patternS from extra chart input field}
}
\value{
Text to be displayed
}
\description{
Text to be displayed in "extra tab" in document box
}
