% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cx_extra_functions.R
\name{cx_indices_one_doc}
\alias{cx_indices_one_doc}
\title{Generate identified chunk indices for one doc, one phrase}
\usage{
cx_indices_one_doc(input_field, case_sensitive, doc_index)
}
\description{
Internal. And works in the Python process
}
