% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cx_extra_functions.R
\name{cx_parse_terms}
\alias{cx_parse_terms}
\title{Parse one input phrase into something useful for further processing}
\usage{
cx_parse_terms(extra_input_phrase, case_sensitive)
}
\arguments{
\item{case_sensitive}{search_arguments$case_sensitive}

\item{input_field}{One pattern collected from input field}
}
\value{
Something useful, e.g. parsed phrase
  "ost--0" to c("ost", "0")
}
\description{
Parse one input phrase into something useful for further processing
}
