% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cx_setup.R
\name{cx_setup}
\alias{cx_setup}
\title{Initiate Python environment}
\usage{
cx_setup(
  virtenv =
    "/Users/kristianlg/Library/CloudStorage/Dropbox/Projects/corporaexplorer_development/cxsentences/inst/python/.venv/"
)
}
\arguments{
\item{virtenv}{Path to virtenv with required modules}
}
\description{
Initiate Python environment
}
