% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cx_extra_functions.R
\name{cx_shiny_validate}
\alias{cx_shiny_validate}
\title{"shiny::validate(shiny::need(" wrapper for cx_validate_input()}
\usage{
cx_shiny_validate(extra_patterns)
}
\arguments{
\item{extra_patterns}{character vector of "extra input phrases"}
}
\description{
"shiny::validate(shiny::need(" wrapper for cx_validate_input()
}
