% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cx_extra_functions.R
\name{cx_validate_input}
\alias{cx_validate_input}
\title{Validate input phrases}
\usage{
cx_validate_input(input_field)
}
\arguments{
\item{input_field}{character vector of "extra input phrases"}
}
\value{
Boolean. TRUE if all phrases are OK.
}
\description{
Validate input phrases
}
