% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence_functions.R
\name{explore_sentence_filtered_df}
\alias{explore_sentence_filtered_df}
\title{runs corporaexplorer `explore0()` on the df returned by
  `create_sentence_filtered_df()`, with two search fields
  pre-populated}
\usage{
explore_sentence_filtered_df(df, filter_pattern, search_pattern, window)
}
\description{
runs corporaexplorer `explore0()` on the df returned by
  `create_sentence_filtered_df()`, with two search fields
  pre-populated
}
