% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence_functions.R
\name{get_filtered_doc_indices_from_py}
\alias{get_filtered_doc_indices_from_py}
\title{Title}
\usage{
get_filtered_doc_indices_from_py(r_indexing = TRUE)
}
\arguments{
\item{r_indexing}{Whether the input is from R index (starting on 1)}
}
\description{
Title
}
