% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence_functions.R
\name{get_sentence_window_indices}
\alias{get_sentence_window_indices}
\title{Wrapper for convenient argument passing}
\usage{
get_sentence_window_indices(filter_pattern, window)
}
\arguments{
\item{filter_pattern}{Chr. around which pattern extract sentences?}

\item{window}{How many sentences to each side of `filter_pattern` to extract?}

\item{search_pattern}{Which pattern to search within extracted sentences?}

\item{r_indexing}{Add 1 to python zero indexing to achieve equivalent
indexing in R?}
}
\value{
R integer vector with (Python zero style) indices for the
  documents that have `search_pattern` in `window` sentence proximity
    to `filter_pattern`.
}
\description{
Important: "silent" argument: `sentences` in the Python process,
  created by `tokenize_sentences_wrapper()`
}
