% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence_functions.R
\name{tokenize_sentences_ru}
\alias{tokenize_sentences_ru}
\title{Tokenize character vector to a list of lists of sentences}
\usage{
tokenize_sentences_ru(df, to_lower = TRUE)
}
\arguments{
\item{df}{R data frame or corporaexplorerobject}

\item{to_lower}{convert the final list of lists to lower case?}
}
\value{
internal Python object `sentences`
  and `sentences_lower`.
}
\description{
IMPORTANT: must be run first in order to create
  internal Python object `sentences`
  and `sentences_lower`
  to be accessed by `get_sentence_filtered_doc_indices()`
  (and thereby also by `create_sentence_filtered_df()`
  and `explore_sentence_filtered_df()`).
}
