% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cx_setup.R
\name{cx_setup}
\alias{cx_setup}
\title{Initiate Python environment}
\usage{
cx_setup(
  virtenv = "/Users/Kristian/.local/share/virtualenvs/apekatt-mr9vg8U1/"
)
}
\arguments{
\item{virtenv}{Path to virtenv with required modules}
}
\value{

}
\description{
Initiate Python environment
}
